refactor(tegg): extract shared message types into AgentMessage.ts
Move ContentBlockType, InputContentPart, TextContentBlock, MessageContentBlock,
InputMessage, and MessageObject into a dedicated AgentMessage.ts file to eliminate
the circular dependency between AgentStore.ts and AgentRuntime.ts (previously
worked around with an inline import() type).
