+++ tegg/core/types/src/agent-runtime/index.ts
@@ -1,3 +1,4 @@
+export * from './AgentMessage.ts';
 export * from './AgentStore.ts';
 export * from './AgentRuntime.ts';
 export * from './ObjectStorageClient.ts';
